refactor: move test from `invalid` option to `UglifyJsPlugin` test
